requirements.txt
# Video Classification Project Dependencies

# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0

# Machine Learning
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.7.0

# Configuration
PyYAML>=6.0

# Video Processing
ffmpeg-python>=0.2.0

# Utilities
tqdm>=4.62.0
pathlib2>=2.3.0

# Optional: Weights & Biases for experiment tracking
wandb>=0.12.0

# Optional: Jupyter for development
jupyter>=1.0.0
ipykernel>=6.0.0
